fix: Cast float damage to int in PlayerService.ChangeHealth call
PlayerService.ChangeHealth expects int parameter, but
CardManager.ComputeDamageAttack returns float. Added explicit cast.

+++ JDG Mobile Game/Assets/_Scripts/Managers/PlayerManager.cs
@@ -81,7 +81,7 @@ public class PlayerManager : Singleton<PlayerManager>
         else
         {
             var diff = CardManager.Instance.ComputeDamageAttack();
-            PlayerService.ChangeHealth(opponentId, diff);
+            PlayerService.ChangeHealth(opponentId, (int)diff);
         }
 
         // Sync legacy PlayerStatus MonoBehaviours with service state
